+++ oshdb-api/src/main/java/org/heigit/ohsome/oshdb/api/db/OSHDBDatabase.java
@@ -1,7 +1,6 @@
 package org.heigit.ohsome.oshdb.api.db;
 
 import java.util.OptionalLong;
-import org.heigit.ohsome.oshdb.OSHDB;
 import org.heigit.ohsome.oshdb.api.mapreducer.MapReducer;
 import org.heigit.ohsome.oshdb.util.exceptions.OSHDBTimeoutException;
 import org.heigit.ohsome.oshdb.util.mappable.OSHDBMapReducible;
@@ -9,7 +8,7 @@ import org.heigit.ohsome.oshdb.util.mappable.OSHDBMapReducible;
 /**
  * OSHDB database backend connector.
  */
-public abstract class OSHDBDatabase extends OSHDB implements AutoCloseable {
+public abstract class OSHDBDatabase implements AutoCloseable {
   private String prefix = "";
   private Long timeout = null;